% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcPctiles}
\alias{calcPctiles}
\title{Determine score percentile positions in a range.}
\usage{
calcPctiles(rangeStart, rangeEnd, targetPos, targetScores, pctiles)
}
\arguments{
\item{rangeStart}{Starting position of the target range.}

\item{rangeEnd}{Ending position of the target range.}

\item{targetPos}{Corresponding position coordinates for the input scores.}

\item{targetScores}{input scores from which to calculate percentile coordinates.}

\item{pctiles}{A numeric vector of percentiles.}
}
\value{
A numeric vector of coordinate percentiles.
}
\description{
Determine score percentile positions in a range.
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
