% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{filterByExpression}
\alias{filterByExpression}
\title{Filter quantification by minimum CPM across samples.}
\usage{
filterByExpression(quant, minCpm = 0.5, minSamples = 3)
}
\arguments{
\item{quant}{A numeric matrix or \code{data.frame} with TSSs in rows and
samples in columns.}

\item{minCpm}{Minimum CPM value to count a sample as expressing.}

\item{minSamples}{Minimum number of samples that must satisfy the CPM
threshold.}
}
\value{
A logical vector of length \code{nrow(quant)} indicating which TSSs pass.
}
\description{
Filter quantification by minimum CPM across samples.
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
