% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidirectional.R
\name{findBidirectional}
\alias{findBidirectional}
\title{Find bidirectional promoters.}
\usage{
findBidirectional(TSS, maxDist = 500)
}
\arguments{
\item{TSS}{A \code{GRanges} object of TSSs.}

\item{maxDist}{The maximum distance between antisense TSSs to consider
them bidirectional.}
}
\value{
An unstranded \code{GRanges} object representing bidirectional promoters.
}
\description{
Find bidirectional promoters.
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
