% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidirectional.R
\name{findDivergent}
\alias{findDivergent}
\title{Find divergent promoters.}
\usage{
findDivergent(pcTSS, ncTSS, maxDist = 500, returnMerged = TRUE)
}
\arguments{
\item{pcTSS}{A \code{GRanges} object containing protein-coding TSSs.}

\item{ncTSS}{A \code{GRanges} object containing non-coding TSSs.}

\item{maxDist}{The maximum distance between antisense non-coding and
protein-coding TSSs allowed for divergent promoters.}

\item{returnMerged}{Whether to return a set of unstranded ranges covering
the divergent promoters, or a list containing separate \code{GRanges} for
divergent protein-coding and non-coding TSSs.}
}
\value{
If \code{returnMerged = TRUE}, a single unstranded \code{GRanges} object,
otherwise a list containing separate \code{GRanges} objects for protein-coding
and non-coding TSSs.
}
\description{
Find divergent promoters.
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
