% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancer.R
\name{quantifyEnhancers}
\alias{quantifyEnhancers}
\title{Sum TSS quantification over enhancer regions.}
\usage{
quantifyEnhancers(enh, TSS, quant, ignoreStrand = TRUE)
}
\arguments{
\item{enh}{A \code{GRanges} of enhancer regions (e.g. from \code{\link[=defineEnhancers]{defineEnhancers()}}).}

\item{TSS}{A \code{GRanges} of TSSs aligned to the rows of \code{quant}.}

\item{quant}{A numeric matrix or \code{data.frame} of TSS quantifications
(rows aligned with \code{TSS}, samples in columns).}

\item{ignoreStrand}{Whether to ignore strand when overlapping enhancers and
TSSs.}
}
\value{
A numeric matrix of dimensions \verb{length(enh) × ncol(quant)}.
}
\description{
Sum TSS quantification over enhancer regions.
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
