% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{readSignal}
\alias{readSignal}
\title{Read paired stranded or single unstranded bigWig/bedGraph files.}
\usage{
readSignal(p, n = NULL, nScoreIsNegative = NULL, seqLvls = NULL,
  windows = NULL)
}
\arguments{
\item{p}{Filename for the bigWig or bedGraph file to read. If stranded, then
this file represents the positive strand.}

\item{n}{Filename for the negative strand bigWig or bedGraph file to read.}

\item{nScoreIsNegative}{If \code{TRUE}, then ensure the negative strand scores
are negative. If \code{FALSE}, ensure they are positive. If \code{NULL}, leave them
as they are in the original file.}

\item{seqLvls}{If desired, ensure the final \code{GRanges} object uses these
seqlevels.}

\item{windows}{If desired, only read signal overlapping ranges within this
\code{GRanges} object.}
}
\value{
A \code{GRanges} object.
}
\description{
Read paired stranded or single unstranded bigWig/bedGraph files.
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
