% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{readWindowsUnstranded}
\alias{readWindowsUnstranded}
\title{Read signal in windows from a \code{GRanges} object or a bedGraph/bigWig file}
\usage{
readWindowsUnstranded(signal, windows, scoreCol = "score",
  trimSignal = c(0, 1), keepTopPctile = 0, rowNorm = FALSE,
  windowAverage = TRUE, signalFactor = 1, ...)
}
\arguments{
\item{signal}{Either a \code{GRanges} object or a filename.}

\item{windows}{A \code{GRanges} object containing equal width ranges.}

\item{scoreCol}{The name of the column containing scores if the input is a \code{GRanges} object.}

\item{trimSignal}{Lower and upper fractions of the score to trim.}

\item{keepTopPctile}{Only keep windows which are within the top percentile value.}

\item{rowNorm}{Normalize window scores between 0 and 1.}

\item{windowAverage}{Average signal values across all windows.}

\item{signalFactor}{Value to multiply scores by.}

\item{...}{Additional arguments for \code{\link[genomation:ScoreMatrix-methods]{genomation::ScoreMatrix()}}.}
}
\value{
If \code{windowAverage = TRUE}, a \code{data.frame}, otherwise a \code{matrix}.
}
\description{
Read signal in windows from a \code{GRanges} object or a bedGraph/bigWig file
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
