% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swapStrand}
\alias{swapStrand}
\title{Swap strands in a GRanges object.}
\usage{
swapStrand(x)
}
\arguments{
\item{x}{A \code{GRanges} object.}
}
\value{
A sorted \code{GRanges} object with swapped strands for each range.
}
\description{
Swap strands in a GRanges object.
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
