% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{writeBEDPE}
\alias{writeBEDPE}
\title{Write a BEDPE file from two \code{GRanges}}
\usage{
writeBEDPE(x, y, f, names = NULL)
}
\arguments{
\item{x}{The first \code{GRanges} object. It must be in the same order as the second.}

\item{y}{The second \code{GRanges} object. It must be in the same order as the first.}

\item{f}{Output filename.}

\item{names}{Names to use for ranges. If \code{NULL}, names of the form
\code{"INT_<i>"} are generated.}
}
\value{
A \code{data.frame} representation of the BEDPE file, invisibly.
}
\description{
Each pair \verb{(x[i], y[i])} is written as a single BEDPE row. The leftmost
interval of the pair is written as \code{chrom1/start1/end1} and the rightmost
as \code{chrom2/start2/end2}.
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
