% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{tssShapeFromBigWig}
\alias{tssShapeFromBigWig}
\title{File-based wrapper around \code{\link[=tssShape]{tssShape()}}.}
\usage{
tssShapeFromBigWig(TSS, bwDir, samples, posSuffix = ".rpm.pos.bw",
  negSuffix = ".rpm.neg.bw", ...)
}
\arguments{
\item{TSS}{A \code{GRanges} of TSSs.}

\item{bwDir}{Directory containing the bigWigs.}

\item{samples}{Character vector of sample names. For each \code{s}, the files
\verb{<bwDir>/<s><posSuffix>} and \verb{<bwDir>/<s><negSuffix>} are read.}

\item{posSuffix, negSuffix}{File-name suffixes for positive- and
negative-strand bigWigs.}

\item{...}{Further arguments passed to \code{\link[=tssShape]{tssShape()}}.}
}
\value{
The list of matrices returned by \code{\link[=tssShape]{tssShape()}}.
}
\description{
Loads paired stranded bigWig files for a set of samples and runs
\code{\link[=tssShape]{tssShape()}} on the resulting signal list.
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@tsl.ac.uk}
}
